refactor: split out overfilled file module, maintain identical API
