+++ projects/fal/src/fal/toolkit/file/file/__init__.py
@@ -1,0 +1,17 @@
+from .repositories import (
+    BUILT_IN_REPOSITORIES,
+    DEFAULT_REPOSITORY,
+    FileRepositoryFactory,
+    get_builtin_repository,
+)
+from .standard import File
+from .zipped import CompressedFile
+
+__all__ = [
+    "BUILT_IN_REPOSITORIES",
+    "DEFAULT_REPOSITORY",
+    "FileRepositoryFactory",
+    "get_builtin_repository",
+    "File",
+    "CompressedFile",
+]
