+++ projects/fal/src/fal/toolkit/file/file/repositories.py
@@ -1,0 +1,35 @@
+from __future__ import annotations
+
+from typing import Callable
+
+from fal.toolkit.file.providers.fal import FalFileRepository, InMemoryRepository
+from fal.toolkit.file.providers.gcp import GoogleStorageRepository
+from fal.toolkit.file.providers.r2 import R2Repository
+from fal.toolkit.file.types import FileRepository, RepositoryId
+
+__all__ = [
+    "FileRepositoryFactory",
+    "BUILT_IN_REPOSITORIES",
+    "get_builtin_repository",
+    "DEFAULT_REPOSITORY",
+]
+
+FileRepositoryFactory = Callable[[], FileRepository]
+
+BUILT_IN_REPOSITORIES: dict[RepositoryId, FileRepositoryFactory] = {
+    "fal": lambda: FalFileRepository(),
+    "in_memory": lambda: InMemoryRepository(),
+    "gcp_storage": lambda: GoogleStorageRepository(),
+    "r2": lambda: R2Repository(),
+}
+
+
+def get_builtin_repository(id: RepositoryId) -> FileRepository:
+    if id not in BUILT_IN_REPOSITORIES.keys():
+        raise ValueError(f'"{id}" is not a valid built-in file repository')
+    return BUILT_IN_REPOSITORIES[id]()
+
+
+get_builtin_repository.__module__ = "__main__"
+
+DEFAULT_REPOSITORY: FileRepository | RepositoryId = "fal"
