+++ projects/fal/src/fal/toolkit/file/file/standard.py
@@ -1,39 +1,18 @@
 from __future__ import annotations
 
 from pathlib import Path
-from tempfile import NamedTemporaryFile, TemporaryDirectory
-from typing import Any, Callable, Optional
+from typing import Any, Optional
 from urllib.parse import urlparse
-from zipfile import ZipFile
 
 from pydantic import BaseModel, Field, PrivateAttr
 
-from fal.toolkit.file.providers.fal import FalFileRepository, InMemoryRepository
-from fal.toolkit.file.providers.gcp import GoogleStorageRepository
-from fal.toolkit.file.providers.r2 import R2Repository
 from fal.toolkit.file.types import FileData, FileRepository, RepositoryId
 from fal.toolkit.mainify import mainify
 from fal.toolkit.utils.download_utils import download_file
 
-FileRepositoryFactory = Callable[[], FileRepository]
+from .repositories import DEFAULT_REPOSITORY, get_builtin_repository
 
-BUILT_IN_REPOSITORIES: dict[RepositoryId, FileRepositoryFactory] = {
-    "fal": lambda: FalFileRepository(),
-    "in_memory": lambda: InMemoryRepository(),
-    "gcp_storage": lambda: GoogleStorageRepository(),
-    "r2": lambda: R2Repository(),
-}
-
-
-def get_builtin_repository(id: RepositoryId) -> FileRepository:
-    if id not in BUILT_IN_REPOSITORIES.keys():
-        raise ValueError(f'"{id}" is not a valid built-in file repository')
-    return BUILT_IN_REPOSITORIES[id]()
-
-
-get_builtin_repository.__module__ = "__main__"
-
-DEFAULT_REPOSITORY: FileRepository | RepositoryId = "fal"
+__all__ = ["File"]
 
 
 @mainify
@@ -150,39 +129,3 @@ class File(BaseModel):
         downloaded_path.rename(file_path)
 
         return file_path
-
-
-@mainify
-class CompressedFile(File):
-    _extract_dir: Optional[TemporaryDirectory] = PrivateAttr(default=None)
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self._extract_dir = None
-
-    def __iter__(self):
-        if not self._extract_dir:
-            self._extract_files()
-
-        files = Path(self._extract_dir.name).iterdir()  # type: ignore
-        return iter(files)
-
-    def _extract_files(self):
-        self._extract_dir = TemporaryDirectory()
-
-        with NamedTemporaryFile() as temp_file:
-            file_path = temp_file.name
-            self.save(file_path, overwrite=True)
-
-            with ZipFile(file_path) as zip_file:
-                zip_file.extractall(self._extract_dir.name)
-
-    def glob(self, pattern: str):
-        if not self._extract_dir:
-            self._extract_files()
-
-        return Path(self._extract_dir.name).glob(pattern)  # type: ignore
-
-    def __del__(self):
-        if self._extract_dir:
-            self._extract_dir.cleanup()
